extend docs and fix "idempotency" issues of wordpress entrypoint container script

+++ wp-entrypoint.sh
@@ -24,15 +24,22 @@ PHP
 wp core install --path="$WP_PATH" --url="http://127.0.0.1:8000" --title="WWF Plugin" --admin_user=admin --admin_password=password --admin_email="test@test.local"
 # remove all plugins
 wp plugin list
-wp plugin deactivate --quiet akismet hello
-wp plugin delete --quiet akismet hello
+
+wp plugin deactivate --quiet akismet hello || true
+wp plugin delete --quiet akismet hello || true
+
 # install and activate woocommerce plugin
 wp plugin install --activate woocommerce
 wp plugin install --activate woocommerce-services
+wp plugin install --activate blackbox-debug-bar
+
 # link donations plugin to wp-content/plugins
-ln -s /var/www/donations-plugin/ /var/www/html/wp-content/plugins/donations-plugin
+if [ ! -L /var/www/html/wp-content/plugins/donations-plugin ]; then
+  echo "creating symlink dir for plugin"
+  ln -s /var/www/donations-plugin/ /var/www/html/wp-content/plugins/donations-plugin
+fi
 # .. and activate it
-wp plugin activate donations-plugin
+wp plugin activate donations-plugin || true
 # set proper theme
 wp theme install --activate shophistic-lite
